# openapi.yaml
openapi: 3.1.0
info:
  title: KORION KORI Backend API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/onchain/networks/{network}/wallets/{address}/balance:
    get:
      summary: Lookup on-chain TRX and KORI balance for any address on mainnet or testnet
      parameters:
        - in: path
          name: network
          required: true
          schema:
            type: string
            enum: [mainnet, testnet]
        - in: path
          name: address
          required: true
          schema:
            type: string
            pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
      responses:
        '200':
          description: On-chain balance lookup response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnchainBalanceResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/wallets/address-binding:
    post:
      summary: Bind a wallet address to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletBindingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  binding:
                    $ref: '#/components/schemas/WalletBinding'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
    get:
      summary: Lookup wallet binding by userId or walletAddress
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: string
        - in: query
          name: walletAddress
          required: false
          schema:
            type: string
            pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
      responses:
        '200':
          description: Wallet binding lookup response
          content:
            application/json:
              schema:
                type: object
                properties:
                  binding:
                    oneOf:
                      - $ref: '#/components/schemas/WalletBinding'
                      - type: 'null'
  /api/wallets/balance:
    get:
      summary: Get wallet balance by userId or walletAddress
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: string
        - in: query
          name: walletAddress
          required: false
          schema:
            type: string
            pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
      responses:
        '200':
          description: Balance response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /api/wallets/timeline:
    get:
      summary: Get wallet timeline by userId or walletAddress
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: string
        - in: query
          name: walletAddress
          required: false
          schema:
            type: string
            pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Wallet timeline response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletTimelineResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /api/wallets/{userId}/balance:
    get:
      summary: Get wallet balance by userId path
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Balance response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
  /api/wallets/{userId}/address:
    get:
      summary: Get bound wallet address by userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wallet address response
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  walletAddress:
                    type: string
                    nullable: true
  /api/wallets/transfer:
    post:
      summary: Internal transfer
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalTransferRequest'
      responses:
        '201':
          description: Created
        '200':
          description: Duplicate request
        '400':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    WithdrawAdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  schemas:
...
    InternalTransferRequest:
      type: object
      required: [amount]
      properties:
        fromUserId:
          type: string
        fromWalletAddress:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        toUserId:
          type: string
        toWalletAddress:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        amount:
          type: number
          minimum: 0
...
    WalletBindingRequest:
      type: object
      required: [userId, walletAddress]
      properties:
        userId:
          type: string
        walletAddress:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
    WalletBinding:
      type: object
      properties:
        userId:
          type: string
        walletAddress:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        createdAt:
          type: string
          format: date-time
...
    WalletMonitoringSnapshot:
      type: object
      properties:
        address:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        tokenSymbol:
          type: string
        tokenContractAddress:
          type: string
          nullable: true
        tokenBalance:
          type: string
          nullable: true
        tokenRawBalance:
          type: string
          nullable: true
        tokenDecimals:
          type: integer
          nullable: true
        trxBalance:
          type: string
          nullable: true
        trxRawBalance:
          type: string
          nullable: true
        fetchedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [ok, error]
        error:
          type: string
          nullable: true
...
    OnchainBalanceResponse:
      type: object
      properties:
        network:
          type: string
          enum: [mainnet, testnet]
        tronApiUrl:
          type: string
        tokenContractAddress:
          type: string
        wallet:
          $ref: '#/components/schemas/WalletMonitoringSnapshot'
...
    BalanceResponse:
      type: object
      properties:
        userId:
          type: string
        walletAddress:
          type: string
          nullable: true
        balance:
          type: string
        lockedBalance:
          type: string
        updatedAt:
          type: string
          format: date-time
    WalletTimelineEntry:
      type: object
      properties:
        timelineId:
          type: string
        entryType:
          type: string
          enum: [deposit, withdrawal, internal_transfer_in, internal_transfer_out]
        amount:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        txHash:
          type: string
          nullable: true
        blockNumber:
          type: integer
          nullable: true
        counterpartyUserId:
          type: string
          nullable: true
        toAddress:
          type: string
          nullable: true
        depositId:
          type: string
          nullable: true
        withdrawalId:
          type: string
          nullable: true
        ledgerTxId:
          type: string
          nullable: true
        displayStatus:
          type: string
          enum: [submitted, under_review, approved, sending, completed, failed, rejected]
          nullable: true
    WalletTimelineResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WalletTimelineEntry'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true
...